"Map availability field in MentorResource, handle string-based availability data"
This commit updates the `toArray` method in the `MentorResource` class to modify how the `availability` field is mapped. If the availability data is a string, it will decode the JSON data and return a simplified array with 'date' and 'time\_slots' keys. If the data is not a string or cannot be decoded as JSON, it will be returned as is. Previously, the code attempted to map the availability data using a more complex structure, but this has been simplified in this commit.